# homework/Anti-plagiarism-service/api-files/openapi.yaml
openapi: 3.0.3
info:
  title: Anti-Plagiarism Service API
  version: 1.0.0
  description: Anti-plagiarism service for managing student works, submissions, and
    verification reports.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: http://158.160.186.61:8080/api/v1
    description: Public
...
paths:
...
  /works/{workId}/reports:
    get:
      tags:
        - Reports
      summary: Get analytics by workId
      description: Get a list of reports/analytics for all submissions associated
        with the given workId.
      x-codeSamples:
        - lang: cURL
          label: Fetch reports
          source: 'curl http://localhost:8080/works/hw-kpo-3/reports

          '
      parameters:
        - name: workId
          in: path
          required: true
          description: Work identifier
          schema:
            type: string
          example: hw-kpo-3
      responses:
        '200':
          description: List of reports for the work
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkReportsResponse'
              examples:
                mixedStatuses:
                  summary: Mixed DONE and PROCESSING reports
                  value:
                    - reportId: rep-001
                      submissionId: sub-001
                      status: DONE
                      plagiarismDetected: true
                      similarityPercent: 45.5
                      createdAt: '2025-11-26T13:00:00Z'
                    - reportId: rep-002
                      submissionId: sub-002
                      status: PROCESSING
                      plagiarismDetected: false
                      similarityPercent: 0
                      createdAt: '2025-11-26T13:05:00Z'
                allClean:
                  summary: Reports without detected plagiarism
                  value:
                    - reportId: rep-010
                      submissionId: sub-010
                      status: DONE
                      plagiarismDetected: false
                      similarityPercent: 3.4
                      createdAt: '2025-11-27T09:15:00Z'
                    - reportId: rep-011
                      submissionId: sub-011
                      status: DONE
                      plagiarismDetected: false
                      similarityPercent: 1.1
                      createdAt: '2025-11-27T09:20:00Z'
        '404':
          description: Work not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: WORK_NOT_FOUND
                message: Work with this workId not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: Unexpected server error
      operationId: listWorkReports
  /submissions/{submissionId}:
    get:
      tags:
        - Submissions
      summary: Get submission details
      description: Get detailed information about a specific work submission.
      x-codeSamples:
        - lang: cURL
          label: Fetch submission
          source: 'curl http://localhost:8080/submissions/sub-001

          '
      parameters:
        - name: submissionId
          in: path
          required: true
          description: Submission identifier
          schema:
            type: string
          example: sub-001
      responses:
        '200':
          description: Submission details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
              examples:
                done:
                  summary: Completed report available
                  value:
                    submissionId: sub-001
                    workId: hw-kpo-3
                    fileId: f47ac10b-58cc-4372-a567-0e02b2c3d479
                    status: DONE
                    uploadedAt: '2025-11-26T12:30:00Z'
                    report:
                      status: DONE
                      plagiarismDetected: true
                      similarityPercent: 45.5
                      matchedSubmissions:
                        - submissionId: sub-0002
                          matchedChunks: 12
                          similarityPercent: 78.0
                      reportId: rep-001
                      createdAt: '2025-11-26T12:35:00Z'
                pending:
                  summary: Pending plagiarism check
                  value:
                    submissionId: sub-099
                    workId: hw-kpo-3
                    fileId: c56a4180-65aa-42ec-a945-5fd21dec0538
                    status: QUEUED
                    uploadedAt: '2025-11-27T10:10:00Z'
        '404':
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: SUBMISSION_NOT_FOUND
                message: Submission with this submissionId not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: Unexpected server error
      operationId: getSubmission
...
components:
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - WORK_ALREADY_EXISTS
        - WORK_NOT_FOUND
        - SUBMISSION_NOT_FOUND
        - REPORT_NOT_FOUND
        - INTERNAL_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Name of the field that caused the error
        message:
          type: string
          description: Human-readable description of the issue
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        details:
          type: array
          description: Optional list of validation errors
          items:
            $ref: '#/components/schemas/ErrorDetail'
...
    SubmissionDetail:
      type: object
      required:
        - submissionId
        - workId
        - fileId
        - status
        - uploadedAt
      properties:
        submissionId:
          type: string
          description: Unique submission identifier
          example: sub-001
        workId:
          type: string
          description: Work identifier
          example: hw-kpo-3
        fileId:
          type: string
          format: uuid
          description: File ID in file-storing service
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        status:
          type: string
          enum:
            - QUEUED
            - PROCESSING
            - DONE
            - ERROR
          description: Current status of plagiarism check
          example: DONE
        uploadedAt:
          type: string
          format: date-time
          description: Timestamp when submission was uploaded
          example: '2025-12-11T12:30:00Z'
        report:
          $ref: '#/components/schemas/PlagiarismReport'
    PlagiarismReport:
      type: object
      required:
        - reportId
        - status
        - plagiarismDetected
        - similarityPercent
        - createdAt
      properties:
        reportId:
          type: string
          description: Unique report identifier
          example: rep-001
        status:
          type: string
          enum:
            - QUEUED
            - PROCESSING
            - DONE
            - ERROR
          description: Report status
          example: DONE
        plagiarismDetected:
          type: boolean
          description: Whether plagiarism was detected (typically if similarityPercent
            > 50%)
          example: true
        similarityPercent:
          type: number
          format: float
          description: Overall similarity percentage (0-100)
          minimum: 0
          maximum: 100
          example: 78.5
        createdAt:
          type: string
          format: date-time
          description: Timestamp when report was created
          example: '2025-12-11T13:00:00Z'
        completedAt:
          type: string
          format: date-time
          description: Timestamp when analysis completed (only if status is DONE)
          example: '2025-12-11T13:05:00Z'
        errorMessage:
          type: string
          description: Error message if status is ERROR
          example: Failed to extract text from document
        matchedSubmissions:
          type: array
          description: List of submissions with high similarity
          items:
            $ref: '#/components/schemas/MatchedSubmission'
    MatchedSubmission:
      type: object
      required:
        - submissionId
        - similarityPercent
        - matchedChunks
      properties:
        submissionId:
          type: string
          description: ID of the matched submission
          example: sub-042
        similarityPercent:
          type: number
          format: float
          description: Similarity percentage with this submission
          minimum: 0
          maximum: 100
          example: 78.5
        matchedChunks:
          type: integer
          description: Number of text chunks that matched
          minimum: 0
          example: 12
    WorkReportItem:
      type: object
      required:
        - reportId
        - submissionId
        - status
        - plagiarismDetected
        - similarityPercent
        - createdAt
      properties:
        reportId:
          type: string
          description: Unique report identifier
          example: rep-001
        submissionId:
          type: string
          description: Submission identifier
          example: sub-001
        status:
          type: string
          enum:
            - QUEUED
            - PROCESSING
            - DONE
            - ERROR
          description: Report status
          example: DONE
        plagiarismDetected:
          type: boolean
          description: Whether plagiarism was detected
          example: true
        similarityPercent:
          type: number
          format: float
          description: Similarity percentage (0-100)
          minimum: 0
          maximum: 100
          example: 78.5
        createdAt:
          type: string
          format: date-time
          description: Timestamp when report was created
          example: '2025-12-11T13:00:00Z'
        completedAt:
          type: string
          format: date-time
          description: Timestamp when analysis completed
          example: '2025-12-11T13:05:00Z'
    WorkReportsResponse:
      type: array
      items:
        $ref: '#/components/schemas/WorkReportItem'
...
security: []