# homework/Anti-plagiarism-service/file-analisys/internal/api/generated/generated.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for AnalyzeResponseStatus.
const (
	AnalyzeResponseStatusQUEUED AnalyzeResponseStatus = "QUEUED"
)

// Defines values for ErrorCode.
const (
	FILENOTFOUND       ErrorCode = "FILE_NOT_FOUND"
	INTERNALERROR      ErrorCode = "INTERNAL_ERROR"
	REPORTNOTFOUND     ErrorCode = "REPORT_NOT_FOUND"
	SERVICEUNAVAILABLE ErrorCode = "SERVICE_UNAVAILABLE"
	VALIDATIONERROR    ErrorCode = "VALIDATION_ERROR"
	WORKNOTFOUND       ErrorCode = "WORK_NOT_FOUND"
)

// Defines values for ReportResponseStatus.
const (
	ReportResponseStatusDONE       ReportResponseStatus = "DONE"
	ReportResponseStatusERROR      ReportResponseStatus = "ERROR"
	ReportResponseStatusPROCESSING ReportResponseStatus = "PROCESSING"
	ReportResponseStatusQUEUED     ReportResponseStatus = "QUEUED"
)

// Defines values for WorkReportItemStatus.
const (
	DONE       WorkReportItemStatus = "DONE"
	ERROR      WorkReportItemStatus = "ERROR"
	PROCESSING WorkReportItemStatus = "PROCESSING"
	QUEUED     WorkReportItemStatus = "QUEUED"
)

// AnalyzeRequest defines model for AnalyzeRequest.
type AnalyzeRequest struct {
	// FileId File ID from file-storing service
	FileId openapi_types.UUID `json:"fileId"`

	// SubmissionId Unique submission identifier
	SubmissionId string `json:"submissionId"`

	// WorkId Work identifier (embeddings are grouped by workId)
	WorkId string `json:"workId"`
}

// AnalyzeResponse defines model for AnalyzeResponse.
type AnalyzeResponse struct {
	// Message Human-readable message
	Message *string `json:"message,omitempty"`

	// QueuedAt Timestamp when queued
	QueuedAt time.Time `json:"queuedAt"`

	// Status Initial status
	Status AnalyzeResponseStatus `json:"status"`

	// SubmissionId Submission identifier
	SubmissionId string `json:"submissionId"`

	// WorkId Work identifier
	WorkId string `json:"workId"`
}

// AnalyzeResponseStatus Initial status
type AnalyzeResponseStatus string

// ErrorCode defines model for ErrorCode.
type ErrorCode string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code ErrorCode `json:"code"`

	// Details Additional error context
	Details *map[string]interface{} `json:"details,omitempty"`

	// Message Human-readable error message
	Message string `json:"message"`
}

// MatchedSubmission defines model for MatchedSubmission.
type MatchedSubmission struct {
	// MatchedChunks Number of chunks that matched
	MatchedChunks int `json:"matchedChunks"`

	// SimilarityPercent Similarity percentage with this submission
	SimilarityPercent float64 `json:"similarityPercent"`

	// SubmissionId ID of matched submission
	SubmissionId string `json:"submissionId"`
}

// ReportResponse defines model for ReportResponse.
type ReportResponse struct {
	// CompletedAt Timestamp when analysis completed (only for DONE status)
	CompletedAt *time.Time `json:"completedAt,omitempty"`

	// CreatedAt Timestamp when report was created
	CreatedAt time.Time `json:"createdAt"`

	// ErrorMessage Error message (only for ERROR status)
	ErrorMessage *string `json:"errorMessage,omitempty"`

	// MatchedSubmissions List of submissions with high similarity
	MatchedSubmissions *[]MatchedSubmission `json:"matchedSubmissions,omitempty"`

	// PlagiarismDetected True if similarityPercent > 50%
	PlagiarismDetected bool `json:"plagiarismDetected"`

	// SimilarityPercent Similarity percentage (0-100)
	SimilarityPercent float32 `json:"similarityPercent"`

	// Status Current analysis status
	Status ReportResponseStatus `json:"status"`

	// SubmissionId Submission identifier
	SubmissionId string `json:"submissionId"`

	// WorkId Work identifier
	WorkId string `json:"workId"`
}

// ReportResponseStatus Current analysis status
type ReportResponseStatus string

// WorkReportItem defines model for WorkReportItem.
type WorkReportItem struct {
	CompletedAt        *time.Time           `json:"completedAt,omitempty"`
	CreatedAt          time.Time            `json:"createdAt"`
	PlagiarismDetected bool                 `json:"plagiarismDetected"`
	SimilarityPercent  float32              `json:"similarityPercent"`
	Status             WorkReportItemStatus `json:"status"`
	SubmissionId       string               `json:"submissionId"`
}

// WorkReportItemStatus defines model for WorkReportItem.Status.
type WorkReportItemStatus string

// WorkReportsResponse defines model for WorkReportsResponse.
type WorkReportsResponse = []WorkReportItem

// AnalyzeFileJSONRequestBody defines body for AnalyzeFile for application/json ContentType.
type AnalyzeFileJSONRequestBody = AnalyzeRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Start plagiarism analysis for a file
	// (POST /analyze)
	AnalyzeFile(w http.ResponseWriter, r *http.Request)
	// Get plagiarism report
	// (GET /reports/{submissionId})
	GetReport(w http.ResponseWriter, r *http.Request, submissionId string)
	// Get all reports for a work
	// (GET /works/{workId}/reports)
	GetWorkReports(w http.ResponseWriter, r *http.Request, workId string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Start plagiarism analysis for a file
// (POST /analyze)
func (_ Unimplemented) AnalyzeFile(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get plagiarism report
// (GET /reports/{submissionId})
func (_ Unimplemented) GetReport(w http.ResponseWriter, r *http.Request, submissionId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get all reports for a work
// (GET /works/{workId}/reports)
func (_ Unimplemented) GetWorkReports(w http.ResponseWriter, r *http.Request, workId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// AnalyzeFile operation middleware
func (siw *ServerInterfaceWrapper) AnalyzeFile(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AnalyzeFile(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetReport operation middleware
func (siw *ServerInterfaceWrapper) GetReport(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "submissionId" -------------
	var submissionId string

	err = runtime.BindStyledParameterWithOptions("simple", "submissionId", chi.URLParam(r, "submissionId"), &submissionId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "submissionId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetReport(w, r, submissionId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetWorkReports operation middleware
func (siw *ServerInterfaceWrapper) GetWorkReports(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "workId" -------------
	var workId string

	err = runtime.BindStyledParameterWithOptions("simple", "workId", chi.URLParam(r, "workId"), &workId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "workId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWorkReports(w, r, workId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/analyze", wrapper.AnalyzeFile)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/reports/{submissionId}", wrapper.GetReport)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/works/{workId}/reports", wrapper.GetWorkReports)
	})

	return r
}